requirements.txt
streamlit
google-generativeai
python-dotenv
openpyxl
seaborn